requirements.txt
diffusers
accelerate
torchvision
transformers>=4.25.1
ftfy
tensorboard
opencv-python
Pillow
pandas
torchaudio
datasets
